♻️ Refactor core functionality and cleanup unused imports
